Move compare button to header in Climate Goals section
