+++ app/javascript/app/constants/links.js
@@ -1,2 +1,3 @@
 export const GLOBAL_CW_PLATFORM = 'https://www.climatewatchdata.org';
 export const WRI_WEBSITE = 'https://www.wri.org/';
+export const COMPARE_NDC_LINK = 'https://www.climatewatchdata.org/ndcs/compare/';
